requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
black==24.8.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.112.0
fastapi-cli==0.0.5
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==2.1.0
openpyxl==3.1.5
orjson==3.9.15
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
pyasn1==0.5.1
pycparser==2.22
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.18.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.20.0
watchfiles==0.21.0
websockets==12.0
XlsxWriter==3.2.0
yarl==1.9.4
